+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/serde/CirceJsonDeserializer.scala
@@ -4,13 +4,17 @@ import cats.data.Validated
 import io.circe
 import io.circe.Json
 import org.everit.json.schema.Schema
-import org.json.{JSONObject, JSONTokener}
+import org.json.JSONTokener
 import pl.touk.nussknacker.engine.api.CirceUtil
-import pl.touk.nussknacker.engine.util.typing.JsonToTypedMapConverter
+import pl.touk.nussknacker.engine.json.SwaggerBasedJsonSchemaTypeDefinitionExtractor
+import pl.touk.nussknacker.engine.json.swagger.SwaggerTyped
+import pl.touk.nussknacker.engine.json.swagger.extractor.JsonToObject
 
 import java.nio.charset.StandardCharsets
 
 class CirceJsonDeserializer(jsonSchema: Schema) {
+
+  val swaggerTyped: SwaggerTyped = SwaggerBasedJsonSchemaTypeDefinitionExtractor.swaggerType(jsonSchema)
 
   def deserialize(bytes: Array[Byte]): Validated[circe.Error, Any] = {
     val string = new String(bytes, StandardCharsets.UTF_8)
@@ -20,6 +24,6 @@ class CirceJsonDeserializer(jsonSchema: Schema) {
   def deserialize(string: String): Validated[circe.Error, Any] = {
     val jsonObject = new JSONTokener(string).nextValue()
     jsonSchema.validate(jsonObject)
-    Validated.fromEither(CirceUtil.decodeJson[Json](string).map(JsonToTypedMapConverter.jsonToAny))
+    Validated.fromEither(CirceUtil.decodeJson[Json](string).map(JsonToObject.apply(_, swaggerTyped)))
   }
 }