+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/swagger/SwaggerTyped.scala
@@ -1,17 +1,17 @@
-package pl.touk.nussknacker.openapi
+package pl.touk.nussknacker.engine.json.swagger
 
 import io.circe.generic.JsonCodec
+import io.swagger.v3.oas.models.media.{ArraySchema, MapSchema, ObjectSchema, Schema}
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedNull, TypedObjectTypingResult, TypingResult}
+import pl.touk.nussknacker.engine.json.swagger.parser.{PropertyName, SwaggerRefSchemas}
 
 import java.time.LocalDateTime
-import io.swagger.v3.oas.models.media.{ArraySchema, MapSchema, ObjectSchema, Schema}
-import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, Typed, TypedClass, TypedObjectTypingResult}
-import pl.touk.nussknacker.openapi.parser.SwaggerRefSchemas
-
 import scala.annotation.tailrec
 import scala.collection.JavaConverters._
 
-@JsonCodec sealed trait SwaggerTyped { self =>
-  def typingResult: SingleTypingResult =
+@JsonCodec sealed trait SwaggerTyped {
+  self =>
+  def typingResult: TypingResult =
     SwaggerTyped.typingResult(self)
 }
 
@@ -22,6 +22,8 @@ case object SwaggerBool extends SwaggerTyped
 case object SwaggerLong extends SwaggerTyped
 
 case object SwaggerDouble extends SwaggerTyped
+
+case object SwaggerNull extends SwaggerTyped
 
 case object SwaggerBigDecimal extends SwaggerTyped
 
@@ -43,7 +45,8 @@ object SwaggerTyped {
     case _ => Option(schema.get$ref()) match {
       case Some(ref) =>
         SwaggerTyped(swaggerRefSchemas(ref), swaggerRefSchemas)
-      case None => (schema.getType, Option(schema.getFormat)) match {
+      case None => (extractType(schema).get, Option(schema.getFormat)) match {
+        case ("object", _) => SwaggerObject(schema.asInstanceOf[ObjectSchema], swaggerRefSchemas)
         case ("boolean", _) => SwaggerBool
         case ("string", Some("date-time")) => SwaggerDateTime
         case ("string", _) => Option(schema.getEnum) match {
@@ -54,12 +57,18 @@ object SwaggerTyped {
         case ("number", None) => SwaggerBigDecimal
         case ("number", Some("double")) => SwaggerDouble
         case ("number", Some("float")) => SwaggerDouble
+        case (null, None) => SwaggerNull
+        //todo handle unions
         case (typeName, format) => throw new Exception(s"Type $typeName in format: $format, is not supported")
       }
     }
   }
 
-  def typingResult(swaggerTyped: SwaggerTyped): SingleTypingResult = swaggerTyped match {
+  private def extractType(schema: Schema[_]): Option[String] =
+    Option(schema.getType)
+      .orElse(Option(schema.getTypes).map(_.asScala.head))
+
+  def typingResult(swaggerTyped: SwaggerTyped): TypingResult = swaggerTyped match {
     case SwaggerObject(elementType, _) =>
       import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
       TypedObjectTypingResult(elementType.mapValuesNow(typingResult).toList.sortBy(_._1))
@@ -79,6 +88,8 @@ object SwaggerTyped {
       Typed.typedClass[java.math.BigDecimal]
     case SwaggerDateTime =>
       Typed.typedClass[LocalDateTime]
+    case SwaggerNull =>
+      TypedNull
   }
 }
 
@@ -95,8 +106,3 @@ object SwaggerObject {
     )
   }
 }
-
-
-
-
-