+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/swagger/parser/parser.scala
@@ -1,8 +1,12 @@
-package pl.touk.nussknacker.openapi
+package pl.touk.nussknacker.engine.json.swagger
 
 import io.swagger.v3.oas.models.media.Schema
 
 package object parser {
+
+  type PropertyName = String
+
+  type SwaggerRef = String
 
   type SwaggerRefSchemas = Map[SwaggerRef, Schema[_]]